countAngleBeetweenPoints dokonczona, update angle tez (nie wiem czy ostatnia linijka jest okej), moveAnt prawie gotowe ( nie wiem co sie stanie jak nie bedzie sie mogla mrowka ruszyc, skad wezmie nowy kat, taki zeby sie ruszyc?),  w map potrzebna nowa funkcja ( to trzeba sprawdzic, trzeba napisac tam funkcje, ktora zaczalem, ktora zwraca wysokosc int w danym punkcie mapy), w agencie i ogolnie wszedzie locx i locy musi byc w incie imo,

+++ src/main/java/org/app/agent/Agent.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 
 public abstract class Agent {
     private java.util.UUID id;
-    private double locx;
-    private double locy;
+    private int locx;
+    private int locy;
 
     public Agent() {
         id = java.util.UUID.randomUUID();
@@ -16,19 +16,19 @@ public abstract class Agent {
         return id;
     }
 
-    public double getLocx() {
+    public int getLocx() {
         return locx;
     }
 
-    public void setLocx(double locx) {
+    public void setLocx(int locx) {
         this.locx = locx;
     }
 
-    public double getLocy() {
+    public int getLocy() {
         return locy;
     }
 
-    public void setLocy(double locy) {
+    public void setLocy(int locy) {
         this.locy = locy;
     }
 }
